+++ src/HTTP/FCL/SwaggerSchema.hs
@@ -35,7 +35,7 @@ import Language.FCL.Metadata as Metadata
 import Language.FCL.Encoding as Encoding
 import Language.FCL.Analysis as Analysis
 import Language.FCL.Undefinedness as Undefinedness
-import Language.FCL.ReachabilityGraph as Reachability
+import Language.FCL.Reachability.Definitions as Reachability
 import qualified Language.FCL.Duplicate as Dupl
 import Numeric.Lossless.Decimal
 import Numeric.Lossless.Number